chore(mt-info-8-4): new comment fixes

+++ src/interfaces/IArcadeStakingRewards.sol
@@ -37,7 +37,7 @@ interface IArcadeStakingRewards {
 
     function getRewardForDuration() external view returns (uint256);
 
-    function lastTimeRewardApplicable() external view returns (uint32);
+    function lastTimeRewardApplicable() external view returns (uint256);
 
     function rewardPerToken() external view returns (uint256);
 
@@ -76,7 +76,7 @@ interface IArcadeStakingRewards {
 
     function withdraw(uint256 amount, uint256 depositId) external;
 
-    function setRewardsDuration(uint32 _rewardsDuration) external;
+    function setRewardsDuration(uint256 _rewardsDuration) external;
 
     function recoverERC20(address tokenAddress, uint256 tokenAmount) external;
